фикс

+++ iNote/Models/NoteInfo.cs
@@ -10,7 +10,7 @@ namespace iNote.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string Desc { get; set; }
-        public byte Color { get; set; }
+        public int Color { get; set; }
         public string LastChange { get; set; }
     }
 }
